TIP-706-textarea: Rework index integration tests
